Remove duplicate of getDatapageDataV2
